LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/column-scroll-marker-focus-002-expected.txt

First
line
line
A
B
C
D
E
F
after

PASS Column #1
FAIL Column #2 assert_equals: expected Element node <div tabindex="0" id="inlineBlock" style="display:inline-... but got Element node <div tabindex="0" id="c1first">line</div>
FAIL Column #0 assert_equals: expected Element node <div tabindex="0" id="c0" style="height:100px;">First</div> but got Element node <div tabindex="0" id="c1first">line</div>
FAIL Column #4 assert_equals: expected Element node <div tabindex="0" id="flex4" style="width:30%;">D</div> but got Element node <div tabindex="0" id="c1first">line</div>
FAIL Column #3 assert_equals: expected Element node <div tabindex="0" id="flex1" style="width:30%; height:100... but got Element node <div tabindex="0" id="c1first">line</div>
